Add audio playing on-click; use EventBus
Add a ref to the audio element to access it from within the local
script for that component; however, given we want to invoke play/pause
on the audio element from other components (namely Region.vue), we must
implement an EventBus (named AudioBus) to achieve that.

We emit events from within the store module to prevent duplication

One thing I do not like about these changes is that selectRegion now
has default, faked properties (particularly an ID of 1 that is known
from our data) to prevent the UI from dying. This will need factored,
but for now we do not care.