Show numbers beside filter options
- Also introduced lexical this (arrow function) to remove the need for
assigning this when the components are mounted.

Note: there is a lot of overlap between these filters (and likely
others in the future); this would best be abstracted to a single
component that can be used with slots, and would remove the need for
components to communicate as the parent would have the single source of
truth.