Only show edit button if admin
I did not previously account for when a project is added from vuex,
given it would not have the current user as a member. Given they have
joined it from this page, then they’re not admins anyway so we check
that members length is > 1 (only true if they are a member from the
server data) before showing the button.